modify nameserver test code
git-svn-id: http://code.taobao.org/svn/tfs@1999 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-tmp-trunk/tests/nameserver/test_database_helper.cpp
@@ -103,7 +103,7 @@ namespace tfs
         for (; iter != info.end(); ++iter)
         {
           std::ostringstream str;
-          std::vector<std::pair<uint32_t, int32_t> >::const_iterator it = (*iter).family_member_.begin();
+          std::vector<std::pair<uint64_t, int32_t> >::const_iterator it = (*iter).family_member_.begin();
           for (; it != (*iter).family_member_.end(); ++it)
             str << " " << (*it).first << ":" << (*it).second << " ";
           TBSYS_LOG(DEBUG, "family id : %ld, family_aid_info: %d, pair: %s",
