+++ branches/tfs-tmp-trunk/tests/nameserver/test_family_manager.cpp
@@ -67,11 +67,11 @@ namespace tfs
     TEST_F(FamilyManagerTest, MarshallingItem_get)
     {
       const uint64_t SERVER_ID = 0xfffffffff;
-      const uint32_t BLOCK_ID  = 100;
+      const uint64_t BLOCK_ID  = 100;
       FamilyManager::MarshallingItem item;
       memset(&item , 0, sizeof(item));
       EXPECT_EQ(TFS_SUCCESS, item.insert(SERVER_ID, BLOCK_ID));
-      std::pair<uint64_t, uint32_t> out;
+      std::pair<uint64_t, uint64_t> out;
       EXPECT_FALSE(item.get(out, 0xffff, 1));
       EXPECT_TRUE(item.get(out, SERVER_ID, BLOCK_ID));
       EXPECT_EQ(SERVER_ID, out.first);
@@ -82,11 +82,11 @@ namespace tfs
     TEST_F(FamilyManagerTest, MarshallingItem_choose_item_random)
     {
       const uint64_t SERVER_ID = 0xfffffffff;
-      const uint32_t BLOCK_ID  = 100;
+      const uint64_t BLOCK_ID  = 100;
       FamilyManager::MarshallingItem item;
       memset(&item , 0, sizeof(item));
 
-      std::pair<uint64_t, uint32_t> out;
+      std::pair<uint64_t, uint64_t> out;
       EXPECT_EQ(EXIT_MARSHALLING_ITEM_QUEUE_EMPTY, item.choose_item_random(out));
       EXPECT_EQ(0, item.slot_num_);
 
@@ -112,9 +112,9 @@ namespace tfs
       SET_CHECK_MEMBER_NUM(family_aid_info, 2);
       SET_MARSHALLING_TYPE(family_aid_info, 1);
       SET_MASTER_INDEX(family_aid_info, 5);
-      std::pair<uint32_t, int32_t> members[MEMBER_NUM];
-      common::ArrayHelper<std::pair<uint32_t, int32_t> > helper(MEMBER_NUM, members);
-      std::pair<uint32_t, int32_t> item;
+      std::pair<uint64_t, int32_t> members[MEMBER_NUM];
+      common::ArrayHelper<std::pair<uint64_t, int32_t> > helper(MEMBER_NUM, members);
+      std::pair<uint64_t, int32_t> item;
       for (int32_t i = 0; i <MEMBER_NUM; i++)
       {
         helper.push_back(std::make_pair(100 + i, i));
@@ -127,15 +127,15 @@ namespace tfs
       FamilyCollect* family2 = family_manager_.get_(family_id);
       EXPECT_TRUE(NULL != family2);
       EXPECT_TRUE(family == family2);
-      std::pair<uint32_t, int32_t> results[MEMBER_NUM];
-      common::ArrayHelper<std::pair<uint32_t, int32_t> > helper2(MEMBER_NUM, results);
+      std::pair<uint64_t, int32_t> results[MEMBER_NUM];
+      common::ArrayHelper<std::pair<uint64_t, int32_t> > helper2(MEMBER_NUM, results);
       EXPECT_EQ(EXIT_NO_FAMILY, family_manager_.get_members_(helper2, 101));
       EXPECT_EQ(TFS_SUCCESS, family_manager_.get_members_(helper2, family_id));
       EXPECT_EQ(MEMBER_NUM , helper2.get_array_index());
       for (int32_t i = 0; i < MEMBER_NUM; ++i)
       {
-        std::pair<uint32_t, int32_t> item = *helper.at(i);
-        std::pair<uint32_t, int32_t> item2 = *helper2.at(i);
+        std::pair<uint64_t, int32_t> item = *helper.at(i);
+        std::pair<uint64_t, int32_t> item2 = *helper2.at(i);
         EXPECT_EQ(item.first, item2.first);
         EXPECT_EQ(item.second, item2.second);
       }
@@ -160,9 +160,9 @@ namespace tfs
       SET_MASTER_INDEX(family_aid_info, 5);
       for (int32_t j = 0; j < MAX_FAMILY_NUM; ++j)
       {
-        std::pair<uint32_t, int32_t> members[MEMBER_NUM];
-        common::ArrayHelper<std::pair<uint32_t, int32_t> > helper(MEMBER_NUM, members);
-        std::pair<uint32_t, int32_t> item;
+        std::pair<uint64_t, int32_t> members[MEMBER_NUM];
+        common::ArrayHelper<std::pair<uint64_t, int32_t> > helper(MEMBER_NUM, members);
+        std::pair<uint64_t, int32_t> item;
         for (int32_t i = 0; i <MEMBER_NUM; i++)
         {
           helper.push_back(std::make_pair(family_id + i, i));
@@ -188,13 +188,13 @@ namespace tfs
           FamilyCollect* family = *result.at(m);
           EXPECT_TRUE(NULL != family);
           EXPECT_EQ(((k * MAX_QUERY_NUM + m) + family_id), family->get_family_id());
-          std::pair<uint32_t, int32_t> results2[MEMBER_NUM];
-          common::ArrayHelper<std::pair<uint32_t, int32_t> > helper2(MEMBER_NUM, results2);
+          std::pair<uint64_t, int32_t> results2[MEMBER_NUM];
+          common::ArrayHelper<std::pair<uint64_t, int32_t> > helper2(MEMBER_NUM, results2);
           family->get_members(helper2);
           for (int32_t n = 0; n < helper2.get_array_index(); ++n)
           {
-            std::pair<uint32_t, int32_t> item2 = *helper2.at(n);
-            EXPECT_EQ((family_id + n), item2.first);
+            std::pair<uint64_t, int32_t> item2 = *helper2.at(n);
+            EXPECT_EQ((uint64_t)(family_id + n), item2.first);
             EXPECT_EQ(n, item2.second);
           }
         }
@@ -204,8 +204,8 @@ namespace tfs
 
     TEST_F(FamilyManagerTest, create_family_choose_data_members)
     {
-      std::pair<uint64_t, uint32_t> members[MAX_MARSHALLING_NUM];
-      common::ArrayHelper<std::pair<uint64_t, uint32_t> > helper(MAX_MARSHALLING_NUM, members);
+      std::pair<uint64_t, uint64_t> members[MAX_MARSHALLING_NUM];
+      common::ArrayHelper<std::pair<uint64_t, uint64_t> > helper(MAX_MARSHALLING_NUM, members);
       int32_t data_member = MAX_DATA_MEMBER_NUM;
       EXPECT_EQ(TFS_SUCCESS, family_manager_.create_family_choose_data_members(helper, data_member));
       EXPECT_EQ(0, helper.get_array_index());
@@ -230,10 +230,10 @@ namespace tfs
       EXPECT_EQ(data_member, helper.get_array_index());
       for (int32_t i = 0; i < helper.get_array_index(); ++i)
       {
-        std::pair<uint64_t, uint32_t> item = *helper.at(i);
+        std::pair<uint64_t, uint64_t> item = *helper.at(i);
         for (int32_t j = i + 1; j < helper.get_array_index(); j++)
         {
-          std::pair<uint64_t, uint32_t> item2 = *helper.at(j);
+          std::pair<uint64_t, uint64_t> item2 = *helper.at(j);
           EXPECT_TRUE(item.first != item2.first);
           EXPECT_TRUE(item.second != item2.second);
         }
